Add more Unit Testing
Command classes do not have unit tests
for execute method.

Adding unit tests to the execute method
helps to validate the response that is
parsed.

Let's add unit tests for execute method
using invalid and valid inputs.

Using a combination of inputs ensures
that the execute method sets the correct
response and throw the correct error