Refactor Command Method
DeleteCommand#execute violates SLAP
and Single Responsibility Principle
because the execute method does the
deletion.

Deletion is done by calling TaskList#
delete(index) and it does a soft delete
by replacing the Task with a flag.
TaskList#wipe does the removing of flags.

Let's implement another method called
deleteSpecificTasks so that an array of
indices can be passed into the method.

The entire deletion process is now handled
by the TaskList object instead of the
Command object.