Add validateIndices method
TaskList#unmarkTasks, markTasks and
deleteSpecificedTasks all contain
duplicate codes to check if their
indices are valid. This means that we
the same code is being run over three
methods

This violates KISS because methods will
be longer than needed.

Let's create a method to validate the
indices so that they can be called in each
of the three methods

By implementing a separate method, our three
methods are shorter and adheres to KISS.