Add validateDates method
Constructor for Event is very long
and may not adhere to KISS and SLAP.

It does not adhere to SLAP because it
does the validation of startTime and
endTime before setting these fields.

Let's implement a method that helps to
validate the arguments passed into the
constructor.

The method validateDates will check if
the two input argument strings are valid.
In this way, we can simply call validateDates
in the constructor to make sure the method of
the constructor remains short.